- OAuth2 dep + HTTPCookie storage and validation
Removed our internal authentication for fastapi from that tutorial we followed from the docs, and instead opted to send the JWT from fastapi to be stored in an HTTP cookie inside frontend, which it then sends back with any request to be validated.  Added a current user dependency, userDep, to add authentication to routes and return the current user.